ScriptChanges
Added Edge GameObject to manage weight.
Added Pi temp element for nodes. Will be useful later on in specific algorithms.
In the making of Graph manager to control the entire graph and be able to run algorithms on it.
Changes to default code working with parameters instead of constant deep hierarchy calls.